use a SandboxUsers group for ACLs instead of granting to each sandbox user.
